Give the remaining text-op constant families one home each
The rest of B14-2's deferred list, ahead of B15's four new sweeps: a
declaration that moves after a sweep is baselined stales its anchors, so
all of this has to land in the prelude.

  * `OP_*` -> a new `types/ops.rs` (28 names / 49 copies / 6 files -> 30 /
    30 / 1)
  * `PUT_*` -> `types/register.rs` (7 / 26 / 6)
  * `CONV_*` and `MB_MAXBYTES` -> `types/mbyte.rs` (6 / 15 / 11 and
    1 / 15 / 15)
  * `BS_*` and `MAX_MCO` -> `types/option.rs`
  * `INSCHAR_*` -> a new `types/edit.rs`

Three families the survey listed turned out to have nothing to
consolidate, and one of them is a trap rather than a no-op:

  * `DIFF_*` is **two unrelated enums sharing a prefix**. `DIFF_NONE` is
    `diffstyle_T = 2` in `diff.rs` and `u8 = 0` in
    `spellsuggest/walk.rs`; consolidating on `^DIFF_` would have merged
    them silently. Left alone — 23 names over 2 files with one collision
    and no duplicate.
  * `REG_*` is `regexp.rs`'s paren classes plus one unrelated
    `REG_TOFREE_KEEP`; five names, five declarations, nothing duplicated.
  * `OP_NONE`/`OP_ADDING`/`OP_PREPENDING`/`OP_REMOVING` are `set_op_T`
    (`:set opt+=`), not operator ids, and are excluded by pattern.

Types are the transpiled ones (`c_uint` for the flag enums), because
retyping a family to its honest type is clippy *growth* until the casts
go with it — that is the rewrite slices' job. The one exception is
`MB_MAXBYTES`, which eight already-rewritten modules use as an array
length: it is `usize`, and the four transpiled comparisons cast instead.
Two `as usize`/`as size_t` casts became same-type casts that way and were
deleted rather than left for clippy to find.

`types/ops.rs` also declares `OP_JOIN_NS` and `OP_FORMAT2`, which never
reached the transpiled tree because nothing named them: `ops.rs` spells
them as the bare numbers 14 and 26 in a `match`. B15-20/21 will want
them.

Seventeen `pub type C2Rust_Unnamed_NN` aliases lost their last user and
were deleted. They were identified by running `prunepreamble.py` against
this rev *and* against the previous one and diffing, rather than by
trusting a single run — the tree has ~230 other prunable preamble items
that predate this slice, and `types/cmdidx.rs`'s 197 are deliberate.

+++ crates/nvim/src/nvim/cmdexpand/pum.rs
@@ -10,6 +10,7 @@
 
 #[allow(unused_imports)]
 use super::*;
+use crate::src::nvim::types::MB_MAXBYTES;
 use core::ffi::{c_char, c_int, c_uint, c_void};
 use core::mem::size_of;
 use core::ptr;
@@ -222,7 +223,7 @@ pub(crate) unsafe fn redraw_wildmenu(
         let mut i;
         let mut l;
 
-        let buf = xmalloc(Columns.get() as size_t * MB_MAXBYTES as size_t + 1) as *mut c_char;
+        let buf = xmalloc(Columns.get() as size_t * MB_MAXBYTES + 1) as *mut c_char;
 
         let mut match_idx = match_idx;
         if match_idx == -1 {
